refactor(portal): rename data access hooks for better clarity
