+++ apps/mini-booking-portal/src/data-access/booking-api/client/use-get-all-rooms.ts
